Harden residual memory-safety risks with modern-C++ guards
Audited every first-party file via a 50-batch parallel review + adversarial
verify, then ran a 5-lens adversarial review of the resulting diff and closed
the sibling sites it flagged. This fixes the focused set of latent
memory-safety defects that survived earlier audit passes. Every change is a
no-op on valid data: byte-exact gameplay parity, ci-test (43/43), ASan+UBSan,
TSan, the WASM build, and the coverage gate all stay green. Adds three
regression tests.

Real defects fixed:

- net_transport read_string: on wasm32 (32-bit size_t) an attacker-controlled
  uint32 length made `offset_ + size` wrap and slip past the bounds check
  (OOB read). Use the overflow-safe subtractive comparison.
- og_file read_uint: signed-overflow UB accumulating into a 32-bit `long` on
  wasm32 before the >1e9 guard could fire. Use a 64-bit accumulator.
- walker/pixien next_frame: modulo-by-zero UB for a 0-frame
  (default/moved-from/test-built) entity; guard both twins.
- walker create_weapon: dereferenced add_ob() without the null check its own
  callers already expect; return nullptr on allocation failure at both sites.
- walker death (life gem), family_slime (x3), family_thief,
  weapon_family_knife, treasure teleporter/life-gem: add_ob() can return null
  (invalid graphics / OOM); guard before dereference, matching the prevalent
  sibling idiom (effect_family_bomb.cpp). Brings the half-guarded call sites
  into line.
- picker_common hire()/accept(): an unvalidated save-loaded team index drove
  an OOB read/write of m_totalcash[4]; clamp at the index sites.
- video_sdl walkputbuffer: the last un-migrated raw videobuffer[] write (OOB
  into the 64000-byte buffer); route through bounds-checked pointb like every
  sibling path.
- level_runtime get_description_line: *end() dereference for a negative index
  on an empty description list; guard the empty case.

Defense-in-depth (no-op today; harden against malformed/future input):

- statistics special_cost/set_special_cost: bounds-check the index.
- button init_buttons: clamp the write loop to array capacity.

Tests: empty-list get_description_line; special_cost out-of-range; next_frame
on a 0-frame walker.

+++ src/resources/io/og_file.cpp
@@ -575,7 +575,11 @@ private:
     {
         if (pos_ >= text_.size() || !std::isdigit(static_cast<unsigned char>(text_[pos_])))
             return fail("expected non-negative integer");
-        long v = 0;
+        // Use a 64-bit accumulator: on ILP32 targets (wasm32/Emscripten) `long`
+        // is 32-bit, so `v * 10` overflowed (signed UB) before the guard below
+        // could reject it. `long long` is >=64-bit everywhere; values that pass
+        // the <=1e9 guard still fit losslessly in the int result.
+        long long v = 0;
         while (pos_ < text_.size() && std::isdigit(static_cast<unsigned char>(text_[pos_]))) {
             v = v * 10 + (text_[pos_] - '0');
             if (v > 1'000'000'000L) return fail("integer overflow");
